auto-install: correctly handle IPv6 addresses
The auto-install does not properly detect IPv6 address when they
are specified in an URL.  Typically, the first IPv6 address part
is to be considered as the hostname and, if specified, a domain
name is appended.  For example,

    http://[fd00:9:152:48:1822:ffff:162:199]/dir/preseed.cfg

becomes

    http://[fd00.example.org:9:152:48:1822:ffff:162:199]/dir/preseed.cfg

which is no longer a valid IPv6 address.  To solve this problem,
enhance auto-install.sh and test for IPv6 addresses in URLs.
Also added few IPv6 unit test cases.

Signed-off-by: Hendrik Brueckner <brueckner@linux.vnet.ibm.com>